+++ OficinaESTG/backend/controllers/PessoaController.php
@@ -8,6 +8,7 @@ use Yii;
 use common\models\Pessoa;
 use yii\data\ActiveDataProvider;
 use yii\web\Controller;
+use yii\web\ForbiddenHttpException;
 use yii\web\NotFoundHttpException;
 use yii\filters\VerbFilter;
 
@@ -37,13 +38,17 @@ class PessoaController extends Controller
      */
     public function actionIndex()
     {
-        $dataProvider = new ActiveDataProvider([
-            'query' => Pessoa::find(),
-        ]);
+        if (\Yii::$app->user->can('viewPessoa')) {
+            $dataProvider = new ActiveDataProvider([
+                'query' => Pessoa::find(),
+            ]);
 
-        return $this->render('index', [
-            'dataProvider' => $dataProvider,
-        ]);
+            return $this->render('index', [
+                'dataProvider' => $dataProvider,
+            ]);
+        }else{
+            throw new ForbiddenHttpException('Não tem permissões', 403);
+        }
     }
 
     /**
@@ -54,14 +59,18 @@ class PessoaController extends Controller
      */
     public function actionView($id)
     {
-        $model =$this->findModel($id);
+        if (\Yii::$app->user->can('viewPessoa')) {
+            $model = $this->findModel($id);
 
-        $carro = Carro::find()->where(['fk_idPessoa'=> $model->idPessoa])->all();
+            $carro = Carro::find()->where(['fk_idPessoa' => $model->idPessoa])->all();
 
-        return $this->render('view', [
-            'model' => $model,
-            'carro' => $carro,
-        ]);
+            return $this->render('view', [
+                'model' => $model,
+                'carro' => $carro,
+            ]);
+        }else{
+            throw new ForbiddenHttpException('Não tem permissões', 403);
+        }
     }
 
     /**
@@ -71,15 +80,19 @@ class PessoaController extends Controller
      */
     public function actionCreate()
     {
-        $model = new Pessoa();
+        if (\Yii::$app->user->can('createPessoa')) {
+            $model = new Pessoa();
 
-        if ($model->load(Yii::$app->request->post()) && $model->save()) {
-            return $this->redirect(['view', 'id' => $model->idPessoa]);
+            if ($model->load(Yii::$app->request->post()) && $model->save()) {
+                return $this->redirect(['view', 'id' => $model->idPessoa]);
+            }
+
+            return $this->render('create', [
+                'model' => $model,
+            ]);
+        }else{
+            throw new ForbiddenHttpException('Não tem permissões', 403);
         }
-
-        return $this->render('create', [
-            'model' => $model,
-        ]);
     }
 
     /**
@@ -91,18 +104,29 @@ class PessoaController extends Controller
      */
     public function actionUpdate($id)
     {
-        $model = $this->findModel($id);
+        if (\Yii::$app->user->can('updatePessoa')) {
+            $model = $this->findModel($id);
 
-        if ($model->load(Yii::$app->request->post())) {
+            if ($model->load(Yii::$app->request->post())) {
 
-            if ($model->save() ){
-                return $this->redirect(['view', 'id' => $model->idPessoa]);
+                if ($model->save()) {
+                    $user_id = User::find()->where(['id' => Yii::$app->user->identity->getId()])->one();
+
+                    $auth = \Yii::$app->authManager;
+                    $utenteRole = $auth->getRole($model->tipoPessoa);
+                    $auth->revokeAll($user_id->id);
+                    $auth->assign($utenteRole, Yii::$app->user->identity->getId());
+
+                    return $this->redirect(['view', 'id' => $model->idPessoa]);
+                }
             }
-        }
 
-        return $this->render('update', [
-            'model' => $model,
-        ]);
+            return $this->render('update', [
+                'model' => $model,
+            ]);
+        }else{
+            throw new ForbiddenHttpException('Não tem permissões', 403);
+        }
     }
 
     /**
@@ -114,10 +138,14 @@ class PessoaController extends Controller
      */
     public function actionDelete($id)
     {
-        $this->findModel($id)->delete();
-        User::deleteAll(['id' => $id]);
+        if (\Yii::$app->user->can('deletePessoa')) {
+            $this->findModel($id)->delete();
+            User::deleteAll(['id' => $id]);
 
-        return $this->redirect(['index']);
+            return $this->redirect(['index']);
+        }else{
+            throw new ForbiddenHttpException('Não tem permissões', 403);
+        }
     }
 
     /**
