+++ OficinaESTG/frontend/controllers/CarroController.php
@@ -6,6 +6,7 @@ use Yii;
 use common\models\Carro;
 use yii\data\ActiveDataProvider;
 use yii\web\Controller;
+use yii\web\ForbiddenHttpException;
 use yii\web\NotFoundHttpException;
 use yii\filters\VerbFilter;
 
@@ -35,11 +36,16 @@ class CarroController extends Controller
      */
     public function actionIndex()
     {
-        $model =Carro::find()->where(['fk_idPessoa' => Yii::$app->user->identity->getId()])->all();
-        
-        return $this->render('index', [
-            'model' => $model
-        ]);
+        if (\Yii::$app->user->can('viewCarro')) {
+            $model =Carro::find()->where(['fk_idPessoa' => Yii::$app->user->identity->getId()])->all();
+
+            return $this->render('index', [
+                'model' => $model
+            ]);
+        }else {
+            throw new ForbiddenHttpException('Não tem permissões', 403);
+        }
+
     }
 
     /**
@@ -50,9 +56,13 @@ class CarroController extends Controller
      */
     public function actionView($id)
     {
-        return $this->render('view', [
-            'model' => $this->findModel($id),
-        ]);
+        if (\Yii::$app->user->can('viewCarro')) {
+            return $this->render('view', [
+                'model' => $this->findModel($id),
+            ]);
+        }else {
+            throw new ForbiddenHttpException('Não tem permissões', 403);
+        }
     }
 
     public function actionView_guest($id)
@@ -69,21 +79,26 @@ class CarroController extends Controller
      */
     public function actionCreate()
     {
-        $model = new Carro();
+        if (\Yii::$app->user->can('createCarro')) {
+            $model = new Carro();
 
-        if ($model->load(Yii::$app->request->post())) {
+            if ($model->load(Yii::$app->request->post())) {
 
-            $model->tipoCarro = 'Reparacao';
-            $model->fk_idPessoa = Yii::$app->user->identity->getId();
+                $model->tipoCarro = 'Reparacao';
+                $model->fk_idPessoa = Yii::$app->user->identity->getId();
 
-            if ($model->save()){
-                return $this->redirect(['view', 'id' => $model->idCarro]);
+                if ($model->save()){
+                    return $this->redirect(['view', 'id' => $model->idCarro]);
+                }
             }
+
+            return $this->render('create', [
+                'model' => $model,
+            ]);
+        } else {
+            throw new ForbiddenHttpException('Não tem permissões', 403);
         }
 
-        return $this->render('create', [
-            'model' => $model,
-        ]);
     }
 
     /**
@@ -95,15 +110,19 @@ class CarroController extends Controller
      */
     public function actionUpdate($id)
     {
-        $model = $this->findModel($id);
+        if (\Yii::$app->user->can('updateCarro')) {
+            $model = $this->findModel($id);
 
-        if ($model->load(Yii::$app->request->post()) && $model->save()) {
-            return $this->redirect(['view', 'id' => $model->idCarro]);
+            if ($model->load(Yii::$app->request->post()) && $model->save()) {
+                return $this->redirect(['view', 'id' => $model->idCarro]);
+            }
+
+            return $this->render('update', [
+                'model' => $model,
+            ]);
+        }else{
+            throw new ForbiddenHttpException('Não tem permissões', 403);
         }
-
-        return $this->render('update', [
-            'model' => $model,
-        ]);
     }
 
     /**
@@ -115,9 +134,13 @@ class CarroController extends Controller
      */
     public function actionDelete($id)
     {
-        $this->findModel($id)->delete();
+        if (\Yii::$app->user->can('deleteCarro')) {
+            $this->findModel($id)->delete();
 
-        return $this->redirect(['index']);
+            return $this->redirect(['index']);
+        }else{
+            throw new ForbiddenHttpException('Não tem permissões', 403);
+        }
     }
 
     /**
